Downloads Job Changes:
The CartoDB API v2 allows for shapefiles to be downloaded directly from
a table, so the DownloadJob now uses this functionality instead of
downloading geojson and converting it with ogr2ogr. This has
significantly decreased the amount of time it takes for a download to be
generate.

This also allows us to easily cache the downloads file and serve from
the cache when no recent edits have been made.

The associated views have been changed as well to work with this new
system, with some of the "Islands I've Edited" leftovers removed along
the way after the feature was removed in a previous commit.